requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
fhir.resources==8.1.0
fhir_core==1.1.4
Flask==3.1.1
greenlet==3.2.4
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
python-dotenv==1.0.1     
pytest==8.3.3            
requests==2.32.3  
Flask-Cors==4.0.1      
loguru==0.7.2          
jsonschema==4.23.0


